requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
appdirs==1.4.4
APScheduler==3.8.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.2.0
billiard==3.6.4.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.6
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
djangorestframework==3.12.4
drf-yasg==1.20.0
Fabric3==1.14.post1
fpdf==1.7.2
ibm-cloud-sdk-core==3.9.0
ibm-watson==5.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mssql-django==1.0
multidict==5.1.0
openapi-codec==1.3.2
packaging==20.9
paramiko==2.7.2
Pillow==8.3.2
prompt-toolkit==3.0.19
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.0.1
PyNaCl==1.4.0
pyodbc==4.0.32
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
pytz==2021.1
redis==3.5.3
reportlab==3.6.1
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
Rx==3.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.1.0
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.4
utils==1.0.1
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.48.0
yarl==1.6.3
